projects/capstone/thecrew_full_stack/thecrew-app-backend-api-python/requirements.txt
alembic==1.4.2
attrs==20.2.0
Bootstrap-Flask==1.5
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.2.1
dynaconf==3.1.0
ecdsa==0.14.1
flasgger==0.9.5
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.5.0
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
mistune==0.8.4
ordereddict==1.1
psycopg2==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
urllib3==1.25.10
Werkzeug==1.0.1
yapf==0.30.0
zipp==3.1.0
